[fix] encode URI for sharing
When sharing a URL for a choice with 2 spaces (e.g. "Gymnase de Nyon"), only the first space was replaced (use `.replace()` instead of `.replaceAll()`). Either ways the better way to do it is to use the [encodeURI](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/encodeURI) function, as in this commit.